// admin/client-next/src/fields/types.ts
export type FieldMeta =
  | { fieldType: 'text'; label: string; path: string }
  | { fieldType: 'textarea'; label: string; path: string }
  | { fieldType: 'email'; label: string; path: string }
  | { fieldType: 'url'; label: string; path: string }
  | { fieldType: 'boolean'; label: string; path: string }
  | { fieldType: 'number'; label: string; path: string }
  | { fieldType: 'numberarray'; label: string; path: string }
  | { fieldType: 'password'; label: string; path: string }
  | { fieldType: 'key'; label: string; path: string }
  | { fieldType: 'name'; label: string; path: string }
  | { fieldType: 'code'; label: string; path: string }
  | { fieldType: 'color'; label: string; path: string }
  | {
      fieldType: 'select';
      label: string;
      path: string;
      options: { value: string | number; label: string }[];
      numeric?: boolean;
      emptyOption?: boolean;
    }
  | { fieldType: 'money'; label: string; path: string }
  | { fieldType: 'date'; label: string; path: string }
  | { fieldType: 'datearray'; label: string; path: string }
  | { fieldType: 'datetime'; label: string; path: string }
  | { fieldType: 'markdown'; label: string; path: string }
  | { fieldType: 'location'; label: string; path: string }
  | { fieldType: 'geopoint'; label: string; path: string }
  | { fieldType: 'html'; label: string; path: string }
  | { fieldType: 'localfile'; label: string; path: string }
  | { fieldType: 'localfiles'; label: string; path: string }
  | {
      fieldType: 'cloudinaryimage';
      label: string;
      path: string;
      folder?: string;
      filenameAssetKey?: string;
    }
  | {
      fieldType: 'cloudinaryimages';
      label: string;
      path: string;
      folder?: string;
      filenameAssetKey?: string;
    }
  | {
      fieldType: 'cloudinary';
      label: string;
      path: string;
      multiple?: boolean;
      folder?: string;
    }
  | { fieldType: 'relationship'; label: string; path: string; refList: string; many?: boolean }
  | { fieldType: 'file'; label: string; path: string }
  | { fieldType: 'textarray'; label: string; path: string };
// ...
export interface FieldProps<TValue> {
  fieldName: string;
  label: string;
  value: TValue;
  onChange: (value: TValue) => void;
  isRequired: boolean;
  isReadonly: boolean;
  errors: string[];
  meta: FieldMeta;
}